refactor(obstacle_stop_planner): clean up code (#1823)
* refactor(obstacle_stop_planner): use alias


* refactor(obstacle_stop_planner): use alias


* refactor(obstacle_stop_planner): reparate utils function


* refactor(obstacle_stop_planner): fix clang-tidy


* refactor(obstacle_stop_planner): separate member functions that can be made static as utils


* refactor(obstacle_stop_planner): rename callback functions


* refactor(obstacle_stop_planner): rename publisher and subscriber


* fix(obstacle_stop_planner): fix typo
