+++ planning/obstacle_stop_planner/include/obstacle_stop_planner/planner_data.hpp
@@ -17,7 +17,12 @@
 
 #include <rclcpp/rclcpp.hpp>
 
+#include <autoware_auto_planning_msgs/msg/trajectory.hpp>
 #include <diagnostic_msgs/msg/diagnostic_status.hpp>
+#include <geometry_msgs/msg/pose.hpp>
+
+#include <pcl/point_cloud.h>
+#include <pcl/point_types.h>
 
 #include <map>
 
@@ -25,6 +30,9 @@ namespace motion_planning
 {
 
 using diagnostic_msgs::msg::DiagnosticStatus;
+using geometry_msgs::msg::Pose;
+
+using autoware_auto_planning_msgs::msg::TrajectoryPoint;
 
 struct StopPoint
 {
@@ -210,7 +218,7 @@ struct PlannerData
 {
   DiagnosticStatus stop_reason_diag{};
 
-  geometry_msgs::msg::Pose current_pose{};
+  Pose current_pose{};
 
   pcl::PointXYZ nearest_collision_point;
